Triggers and FireBall Movement updated
-Generic for trigger, so all you have to to add a new type of switch is
type the array size, drag and drop the gameobject you want, give it a
public void SwitchTrigger() method and make it inherit ISwitchTrigger,
then go into SwitchTriggerScript, and add a const string with whatever
the tag you just added the switchtrigger and iSwitchtrigger to, and then
put it in the start if statement, and the request<ScriptType>(const tag
var), then you are done :).
Also put a timer on switches so it doesn't spastically switch from off
and on.

Fireball movement:  click and hold to move freely much like before,